+++ recpack/datasets/recsys_challenge.py
@@ -26,7 +26,6 @@ class RecsysChallenge2015(Dataset):
         Defaults to `data`
     :type path: str, optional
     :param filename: Name of the file, if no name is provided the dataset default will be used if known.
-        If the dataset does not have a default filename, a ValueError will be raised.
     :type filename: str, optional
     :param use_default_filters: Should a default set of filters be initialised? Defaults to True
     :type use_default_filters: bool, optional