+++ CHANGELOG.md
@@ -7,6 +7,10 @@ _Maintainer | Robin Verachtert | robin.verachtert@froomle.com_
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+### bugfix
+* __algorithms__
+    * Fixed bug in GRU4Rec, which failed when used on a GPU after changes in 0.3.5 .
 
 ## [0.3.6]
 ### Bugfixes