+++ CHANGELOG.md
@@ -7,8 +7,7 @@ _Maintainer | Robin Verachtert | robin.verachtert@froomle.com_
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
-## [UNRELEASED]
+## [0.3.5] - ![](https://img.shields.io/date/1676895426.svg?label=2023-02-20)
 
 ### Bugfixes
 * Replaced remaining usage of `np.bool` with `bool` to be compatible with numpy `1.24.1`